Fix country code selection.
Use a fixed-length buffer instead of a NUL-terminated string. Use that to select country code 0, U.S.A., so that we get @, #, [, ], {, and }. The default country code for the BA63 is Germany, which does not support the square and curly brackets. U.K mode substitutes £ for # which is no good for Twitter hashtags.